% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_logit-EM.R
\name{al_check_convergence}
\alias{al_check_convergence}
\title{Check convergence by computing the maximum rate of change}
\usage{
al_check_convergence(params)
}
\description{
Check convergence by computing the maximum rate of change
}
\keyword{internal}
