% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_logit.R
\name{al_estep}
\alias{al_estep}
\title{E-step function to compute the expectation of omega.}
\usage{
al_estep(X, trials, betas)
}
\description{
E-step function to compute the expectation of omega.
}
\keyword{internal}
