% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_logit-EM.R
\name{al_mstep}
\alias{al_mstep}
\title{M-step function to solve quadratic programming.}
\usage{
al_mstep(y, X, trials, omega, opt_input)
}
\description{
M-step function to solve quadratic programming.
}
\keyword{internal}
