% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_logit.R
\name{al_set_opt_input}
\alias{al_set_opt_input}
\title{Define inputs for constrains}
\usage{
al_set_opt_input(p_vec, regularize, lambda = 1)
}
\arguments{
\item{regularize}{A boolena argument.}

\item{pvec}{A vector of length K where each element corresponds to the level of k-th variable.}
}
\description{
Define inputs for constrains
}
