% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_logit.R
\name{anova_logit}
\alias{anova_logit}
\title{Logistic regression with ANOVA style parametrization.}
\usage{
anova_logit(y, X, option = list())
}
\arguments{
\item{y}{a vector of binary responses.}

\item{X}{a matrix of overparametrized}
}
\description{
Logistic regression with ANOVA style parametrization.
}
