% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbsw.R
\name{cbsw_prepare_params}
\alias{cbsw_prepare_params}
\title{Prepare parameter Inputs}
\usage{
cbsw_prepare_params(obj)
}
\arguments{
\item{obj}{An output object from \code{create_input_matrix}.}
}
\description{
Prepare parameter Inputs
}
\keyword{internal}
