% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbsw.R
\name{cbsw_update_eta}
\alias{cbsw_update_eta}
\title{Update eta}
\usage{
cbsw_update_eta(params, const)
}
\arguments{
\item{params}{A list object of parameters.}

\item{A}{A matrix in the constrains.}

\item{lambda}{A scalar or a vector of regularization parameters.}
}
\value{
A vector of η parameter in the objective.
}
\description{
Update eta
}
\keyword{internal}
