% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbsw.R
\name{cbsw_update_w}
\alias{cbsw_update_w}
\title{Update scaled slack variable}
\usage{
cbsw_update_w(params, const)
}
\arguments{
\item{params}{A list of parameter object.}

\item{A}{A matrix in the constraint. Aβ + Bη = c.}

\item{B}{B matrix in the constraint: Aβ + Bη = c.}
}
\description{
Update scaled slack variable
}
\keyword{internal}
