% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emlogit.R
\name{coef_initialize}
\alias{coef_initialize}
\title{Initialize coefficeints randomly from \code{rnorm()}.}
\usage{
coef_initialize(Y, X, control)
}
\description{
Initialize coefficeints randomly from \code{rnorm()}.
}
\keyword{internal}
