% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_logit-helper.R
\name{create_Amat_unreg}
\alias{create_Amat_unreg}
\title{Create A matrix for linear constraints}
\usage{
create_Amat_unreg(pvec)
}
\description{
Create A matrix for linear constraints
}
\keyword{internal}
