% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_logit-helper.R
\name{create_const_unregularized}
\alias{create_const_unregularized}
\title{Create constraint inputs for no-regularization model.}
\usage{
create_const_unregularized(p_vec)
}
\arguments{
\item{pvec}{A vector of variable levels.}
}
\description{
Create constraint inputs for no-regularization model.
}
\keyword{internal}
