% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbsw.R
\name{cvCBSW}
\alias{cvCBSW}
\title{Covariate Balancing Survey Weights for Categorical Variables with CV}
\usage{
cvCBSW(formula, data, option)
}
\description{
Covariate Balancing Survey Weights for Categorical Variables with CV
}
