% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emlogit.R
\name{emlogit}
\alias{emlogit}
\title{EM algorithm for Multinomial Logit}
\usage{
emlogit(Y, X, control = list())
}
\arguments{
\item{Y}{A matrix of multinomial outcome where rows correspond to observations
and columns correspond to choices.}

\item{X}{A matrix of covariates where columns correspond to variables.
The number of rows should match with the number of rows for \code{Y}.}

\item{control}{A list of the following items:
\itemize{
  \item \code{max_iter} A integer value that specifies the maximum iterations
  for the EM algorithm. Defaul value is 200.
  \item \code{tol} A tolerance for assessing the convergence. Default value is 1e-5.
}}
}
\description{
EM algorithm for Multinomial Logit
}
