% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emlogit-summary.R
\name{predict.emlogit}
\alias{predict.emlogit}
\title{Obtain the predicted probability}
\usage{
\method{predict}{emlogit}(obj, newdata = NULL)
}
\arguments{
\item{obj}{An output of \code{emlogit()} funcion.}

\item{newdata}{A matrix of newdata. If not provided, the insample fit is returned.}
}
\description{
Obtain the predicted probability
}
