% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{predict_prob}
\alias{predict_prob}
\title{Predicted probability}
\usage{
predict_prob(X, B)
}
\arguments{
\item{X}{A matrix of covariates.}

\item{B}{A matrix of coefficeints including the reference category.}
}
\description{
Predicted probability
}
\keyword{internal}
