% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emlogit-summary.R
\name{print.summary.emlogit}
\alias{print.summary.emlogit}
\title{Print function}
\usage{
\method{print}{summary.emlogit}(obj)
}
\arguments{
\item{obj}{An output of \code{summary.emlogit()} function.}
}
\description{
Print function
}
