% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emlogit-summary.R
\name{summary.emlogit.est}
\alias{summary.emlogit.est}
\title{Summary function}
\usage{
\method{summary}{emlogit.est}(obj)
}
\arguments{
\item{obj}{An object class of \code{"emlogit.est"}}
}
\description{
Summary function
}
