% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_logit.R
\name{al_mstep}
\alias{al_mstep}
\title{M-step function to solve quadratic programming.}
\usage{
al_mstep(y, X, omega, p_vec)
}
\description{
M-step function to solve quadratic programming.
}
\keyword{internal}
