% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_logit.R
\name{anova_logit}
\alias{anova_logit}
\title{Logistic regression with ANOVA style parametrization.}
\usage{
anova_logit(y, X, trials = NULL, option = list())
}
\arguments{
\item{y}{A vector of binary responses. In the binomial response, this corresponds to "success".}

\item{X}{A matrix of overparametrized}

\item{trials}{A vector of trial counts. In the binary outcome this is left as \code{NULL}.
The lengh of this vector should match the length of \code{y}.}

\item{option}{A list of options. \code{pvec} should be provided.}
}
\description{
Logistic regression with ANOVA style parametrization.
}
