% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emlogit.R
\name{emlogit}
\alias{emlogit}
\title{EM algorithm for Multinomial Logit}
\usage{
emlogit(Y, X, control = list())
}
\arguments{
\item{Y}{A matrix of multinomial outcomes where rows correspond to observations
and columns correspond to choices.}

\item{X}{A matrix of covariates where columns correspond to variables.
The number of rows should match with the number of rows for \code{Y}.}

\item{control}{A list of control parameters, which can contain the following items:
\itemize{
  \item \code{max_iter} A integer value that specifies the maximum iterations
  for the EM algorithm. Defaul value is 200.
  \item \code{tol} A tolerance parameter for assessing the convergence. Default value is 1e-5.
  \item \code{mu0} A vector of prior means. The dimension of this parameter should match with the dimension of \code{X} (i.e., the number of variables).
   The default value is \code{0}.
  \item \code{Z0} A matrix for the prior variance covariance matrix.
     The dimension of this matrix should match with the dimension of \code{X} i.e., the number of variables).
     The default vaue is \code{diag(rep(5, ncol(X)))}
  \item \code{verbose} A boolean argument. If set \code{TRUE}, the function shows the log-posterior for each iteration. Default is \code{FALSE}.
  \item \code{intercept} A boolean argument. When \code{X} already contains the intercept term (i.e., a column of ones), this option should \code{FALSE}.
     Default is \code{TRUE}.
  \item \code{variance} A boolean argument. If \code{FALSE}, \code{emlogit()} skips the variance calculation.
     This is useful when calling \code{emlogit()} from other programs. Default is \code{TRUE}.
  \item \code{initialize} A method for initialization.
      The package currently supports \code{initialize = "random"} (random) or \code{initialize = "ls"} (least square). Default is \code{"ls"}.
  \item \code{initial_value} A matrix of coefficeints used for the inial value where the first column should be zeros for identification. 
             This matrix should be a dimension of K by J-1 where K is the total number of covariates (including the intercept) and J is the total number of chices.
             If supplied \code{initial_value} overwrites the initial value generated by \code{initialize} option.
}}
}
\description{
EM algorithm for Multinomial Logit
}
